+++ database/migrations/2021_09_15_102218_create_page_translations_table.php
@@ -15,7 +15,7 @@ class CreatePageTranslationsTable extends Migration
     {
         Schema::create('page_translations', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('page_id');
+            $table->integer('page_id');
             $table->text('title');
             $table->text('content');
             $table->text('lang');
